[Type] 타입 폴더 추가
